+++ NEWS.md
@@ -8,7 +8,8 @@
 - Add Rcpp 1.0.7 dependency to fix qs issues.
 - Added a message that reminds the user of the caching behavior. The message will be displayed once every 8 hours. (v1.1.0.07)
 - Added `load_pfr_advstats()` which provides pass, rush, rec, def additional data for each week. (v1.1.0.08)
-- Adjusted caching reminder to check if interactive.
+- Adjusted file location of snap count data (v1.1.0.08)
+- Adjusted caching reminder to check if interactive. (v1.1.0.08)
 
 # nflreadr 1.1.0